phase one of refactoring for #2226
