+++ DNX.sln
@@ -91,8 +91,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "FileSystem", "FileSystem", 
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{0661BDE0-6471-49A4-B902-F826D92EB96D}"
 EndProject
-Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "Microsoft.Dnx.Compilation.Caching", "src\Microsoft.Dnx.Compilation.Caching\Microsoft.Dnx.Compilation.Caching.xproj", "{7004FC14-677B-491F-A98B-E44C61733B46}"
-EndProject
 Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "Microsoft.Dnx.Compilation.DesignTime", "src\Microsoft.Dnx.Compilation.DesignTime\Microsoft.Dnx.Compilation.DesignTime.xproj", "{19D39312-A4A4-449D-B53F-653F951A11F8}"
 EndProject
 Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "Microsoft.Dnx.Host.Tests", "test\Microsoft.Dnx.Host.Tests\Microsoft.Dnx.Host.Tests.xproj", "{6F59A0B7-745D-4ADA-A913-6A1AD1AB49C2}"
@@ -110,6 +108,8 @@ EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "dnx.tests", "test\dnx.tests\dnx.tests.vcxproj", "{9387033F-B955-4794-A6FD-DAD700CBB6A5}"
 EndProject
 Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "Microsoft.Dnx.Compilation.CSharp.Abstractions", "src\Microsoft.Dnx.Compilation.CSharp.Abstractions\Microsoft.Dnx.Compilation.CSharp.Abstractions.xproj", "{EF257BFD-5517-48F5-B16D-407DF86F5A6D}"
+EndProject
+Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "Microsoft.Dnx.Compilation", "src\Microsoft.Dnx.Compilation\Microsoft.Dnx.Compilation.xproj", "{587F95E0-0E56-4C37-8530-6B1485BE01AE}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -785,30 +785,6 @@ Global
 		{970449A2-3BBB-4E87-8513-C41F109B1A3E}.Release|x64.Build.0 = Release|Any CPU
 		{970449A2-3BBB-4E87-8513-C41F109B1A3E}.Release|x86.ActiveCfg = Release|Any CPU
 		{970449A2-3BBB-4E87-8513-C41F109B1A3E}.Release|x86.Build.0 = Release|Any CPU
-		{7004FC14-677B-491F-A98B-E44C61733B46}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{7004FC14-677B-491F-A98B-E44C61733B46}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{7004FC14-677B-491F-A98B-E44C61733B46}.Debug|ARM.ActiveCfg = Debug|Any CPU
-		{7004FC14-677B-491F-A98B-E44C61733B46}.Debug|ARM.Build.0 = Debug|Any CPU
-		{7004FC14-677B-491F-A98B-E44C61733B46}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
-		{7004FC14-677B-491F-A98B-E44C61733B46}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
-		{7004FC14-677B-491F-A98B-E44C61733B46}.Debug|Win32.ActiveCfg = Debug|Any CPU
-		{7004FC14-677B-491F-A98B-E44C61733B46}.Debug|Win32.Build.0 = Debug|Any CPU
-		{7004FC14-677B-491F-A98B-E44C61733B46}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{7004FC14-677B-491F-A98B-E44C61733B46}.Debug|x64.Build.0 = Debug|Any CPU
-		{7004FC14-677B-491F-A98B-E44C61733B46}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{7004FC14-677B-491F-A98B-E44C61733B46}.Debug|x86.Build.0 = Debug|Any CPU
-		{7004FC14-677B-491F-A98B-E44C61733B46}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{7004FC14-677B-491F-A98B-E44C61733B46}.Release|Any CPU.Build.0 = Release|Any CPU
-		{7004FC14-677B-491F-A98B-E44C61733B46}.Release|ARM.ActiveCfg = Release|Any CPU
-		{7004FC14-677B-491F-A98B-E44C61733B46}.Release|ARM.Build.0 = Release|Any CPU
-		{7004FC14-677B-491F-A98B-E44C61733B46}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
-		{7004FC14-677B-491F-A98B-E44C61733B46}.Release|Mixed Platforms.Build.0 = Release|Any CPU
-		{7004FC14-677B-491F-A98B-E44C61733B46}.Release|Win32.ActiveCfg = Release|Any CPU
-		{7004FC14-677B-491F-A98B-E44C61733B46}.Release|Win32.Build.0 = Release|Any CPU
-		{7004FC14-677B-491F-A98B-E44C61733B46}.Release|x64.ActiveCfg = Release|Any CPU
-		{7004FC14-677B-491F-A98B-E44C61733B46}.Release|x64.Build.0 = Release|Any CPU
-		{7004FC14-677B-491F-A98B-E44C61733B46}.Release|x86.ActiveCfg = Release|Any CPU
-		{7004FC14-677B-491F-A98B-E44C61733B46}.Release|x86.Build.0 = Release|Any CPU
 		{19D39312-A4A4-449D-B53F-653F951A11F8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{19D39312-A4A4-449D-B53F-653F951A11F8}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{19D39312-A4A4-449D-B53F-653F951A11F8}.Debug|ARM.ActiveCfg = Debug|Any CPU
@@ -1015,6 +991,30 @@ Global
 		{EF257BFD-5517-48F5-B16D-407DF86F5A6D}.Release|x64.Build.0 = Release|Any CPU
 		{EF257BFD-5517-48F5-B16D-407DF86F5A6D}.Release|x86.ActiveCfg = Release|Any CPU
 		{EF257BFD-5517-48F5-B16D-407DF86F5A6D}.Release|x86.Build.0 = Release|Any CPU
+		{587F95E0-0E56-4C37-8530-6B1485BE01AE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{587F95E0-0E56-4C37-8530-6B1485BE01AE}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{587F95E0-0E56-4C37-8530-6B1485BE01AE}.Debug|ARM.ActiveCfg = Debug|Any CPU
+		{587F95E0-0E56-4C37-8530-6B1485BE01AE}.Debug|ARM.Build.0 = Debug|Any CPU
+		{587F95E0-0E56-4C37-8530-6B1485BE01AE}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
+		{587F95E0-0E56-4C37-8530-6B1485BE01AE}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
+		{587F95E0-0E56-4C37-8530-6B1485BE01AE}.Debug|Win32.ActiveCfg = Debug|Any CPU
+		{587F95E0-0E56-4C37-8530-6B1485BE01AE}.Debug|Win32.Build.0 = Debug|Any CPU
+		{587F95E0-0E56-4C37-8530-6B1485BE01AE}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{587F95E0-0E56-4C37-8530-6B1485BE01AE}.Debug|x64.Build.0 = Debug|Any CPU
+		{587F95E0-0E56-4C37-8530-6B1485BE01AE}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{587F95E0-0E56-4C37-8530-6B1485BE01AE}.Debug|x86.Build.0 = Debug|Any CPU
+		{587F95E0-0E56-4C37-8530-6B1485BE01AE}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{587F95E0-0E56-4C37-8530-6B1485BE01AE}.Release|Any CPU.Build.0 = Release|Any CPU
+		{587F95E0-0E56-4C37-8530-6B1485BE01AE}.Release|ARM.ActiveCfg = Release|Any CPU
+		{587F95E0-0E56-4C37-8530-6B1485BE01AE}.Release|ARM.Build.0 = Release|Any CPU
+		{587F95E0-0E56-4C37-8530-6B1485BE01AE}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
+		{587F95E0-0E56-4C37-8530-6B1485BE01AE}.Release|Mixed Platforms.Build.0 = Release|Any CPU
+		{587F95E0-0E56-4C37-8530-6B1485BE01AE}.Release|Win32.ActiveCfg = Release|Any CPU
+		{587F95E0-0E56-4C37-8530-6B1485BE01AE}.Release|Win32.Build.0 = Release|Any CPU
+		{587F95E0-0E56-4C37-8530-6B1485BE01AE}.Release|x64.ActiveCfg = Release|Any CPU
+		{587F95E0-0E56-4C37-8530-6B1485BE01AE}.Release|x64.Build.0 = Release|Any CPU
+		{587F95E0-0E56-4C37-8530-6B1485BE01AE}.Release|x86.ActiveCfg = Release|Any CPU
+		{587F95E0-0E56-4C37-8530-6B1485BE01AE}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -1055,7 +1055,6 @@ Global
 		{970449A2-3BBB-4E87-8513-C41F109B1A3E} = {0661BDE0-6471-49A4-B902-F826D92EB96D}
 		{A95304E0-DA56-45EE-B5DA-BB6B46D83D88} = {D825D676-CEB4-4D32-8049-A6383AD1929B}
 		{0661BDE0-6471-49A4-B902-F826D92EB96D} = {A95304E0-DA56-45EE-B5DA-BB6B46D83D88}
-		{7004FC14-677B-491F-A98B-E44C61733B46} = {AF391791-F4B7-41AC-8F08-9485DAC543C5}
 		{19D39312-A4A4-449D-B53F-653F951A11F8} = {AF391791-F4B7-41AC-8F08-9485DAC543C5}
 		{6F59A0B7-745D-4ADA-A913-6A1AD1AB49C2} = {C43EE429-DE10-4906-BB09-54E6A080948A}
 		{893A7304-50DB-4053-B864-E10BA3FFAC53} = {13ED5001-B871-4BC3-8499-29607F596C7C}
@@ -1065,5 +1064,6 @@ Global
 		{2AF210A9-5BDC-45E8-95DD-07B5A2616493} = {C43EE429-DE10-4906-BB09-54E6A080948A}
 		{9387033F-B955-4794-A6FD-DAD700CBB6A5} = {C43EE429-DE10-4906-BB09-54E6A080948A}
 		{EF257BFD-5517-48F5-B16D-407DF86F5A6D} = {AF391791-F4B7-41AC-8F08-9485DAC543C5}
+		{587F95E0-0E56-4C37-8530-6B1485BE01AE} = {AF391791-F4B7-41AC-8F08-9485DAC543C5}
 	EndGlobalSection
 EndGlobal
