+++ makefile.shade
@@ -70,8 +70,8 @@ var MANAGED_PROJECTS = '${FindAllProjects(
     BOOTSTRAPPER_HOST_NAME,
     BOOTSTRAPPER_CORECLR_MANAGED_NAME,
     "Microsoft.Dnx.ApplicationHost",
+    "Microsoft.Dnx.Compilation",
     "Microsoft.Dnx.Compilation.Abstractions",
-    "Microsoft.Dnx.Compilation.Caching",
     "Microsoft.Dnx.Compilation.CSharp",
     "Microsoft.Dnx.Compilation.CSharp.Abstractions",
     "Microsoft.Dnx.Compilation.CSharp.Common",
