+++ src/Microsoft.Dnx.Compilation.Abstractions/ILibraryExporter.cs
@@ -1,9 +1,7 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-using Microsoft.Dnx.Compilation;
-
-namespace Microsoft.Dnx.Runtime
+namespace Microsoft.Dnx.Compilation
 {
     /// <summary>
     /// Provides access to the complete graph of dependencies for the application.
