+++ src/Microsoft.Dnx.Compilation.CSharp/RoslynProjectCompiler.cs
@@ -1,10 +1,8 @@
-using Microsoft.CodeAnalysis;
 using System;
 using System.Collections.Generic;
-using Microsoft.Dnx.Runtime.Caching;
-using Microsoft.Dnx.Compilation;
-using Microsoft.Dnx.Runtime.Infrastructure;
+using Microsoft.Dnx.Compilation.Caching;
 using Microsoft.Dnx.Runtime;
+using Microsoft.Dnx.Runtime.Infrastructure;
 
 namespace Microsoft.Dnx.Compilation.CSharp
 {
