+++ src/Microsoft.Dnx.Compilation.CSharp/project.json
@@ -3,7 +3,7 @@
     "description": "ASP.NET 5 Roslyn compiler implementation.",
     "compilationOptions": { "define": [ "TRACE" ], "allowUnsafe": true, "warningsAsErrors": true },
     "dependencies": {
-        "Microsoft.Dnx.Compilation.Caching": "1.0.0-*",
+        "Microsoft.Dnx.Compilation": "1.0.0-*",
         "Microsoft.Dnx.Compilation.CSharp.Common": { "version": "1.0.0-*", "type": "build" },
         "Microsoft.Dnx.Compilation.Abstractions": "1.0.0-*",
         "Microsoft.Dnx.Runtime.Sources": { "version": "1.0.0-*", "type": "build" }
