+++ src/Microsoft.Dnx.Compilation/Caching/FileWriteTimeCacheDependency.cs
@@ -1,8 +1,7 @@
 ﻿using System;
 using System.IO;
-using Microsoft.Dnx.Compilation;
 
-namespace Microsoft.Dnx.Runtime.Caching
+namespace Microsoft.Dnx.Compilation.Caching
 {
     public class FileWriteTimeCacheDependency : ICacheDependency
     {
