+++ src/Microsoft.Dnx.Compilation/Caching/NamedCacheDependency.cs
@@ -1,4 +1,4 @@
-﻿namespace Microsoft.Dnx.Runtime.Caching
+﻿namespace Microsoft.Dnx.Compilation.Caching
 {
     public class NamedCacheDependency : ICacheDependency
     {
