+++ src/Microsoft.Dnx.Compilation/Microsoft.Dnx.Compilation.xproj
@@ -6,10 +6,12 @@
   </PropertyGroup>
   <Import Project="$(VSToolsPath)\DNX\Microsoft.DNX.Props" Condition="'$(VSToolsPath)' != ''" />
   <PropertyGroup Label="Globals">
-    <ProjectGuid>7004fc14-677b-491f-a98b-e44c61733b46</ProjectGuid>
+    <ProjectGuid>587f95e0-0e56-4c37-8530-6b1485be01ae</ProjectGuid>
+    <RootNamespace>Microsoft.Dnx.Compilation</RootNamespace>
     <BaseIntermediateOutputPath Condition="'$(BaseIntermediateOutputPath)'=='' ">..\..\artifacts\obj\$(MSBuildProjectName)</BaseIntermediateOutputPath>
     <OutputPath Condition="'$(OutputPath)'=='' ">..\..\artifacts\bin\$(MSBuildProjectName)\</OutputPath>
   </PropertyGroup>
+
   <PropertyGroup>
     <SchemaVersion>2.0</SchemaVersion>
   </PropertyGroup>
