+++ src/Microsoft.Dnx.Compilation/project.json
