+++ src/Microsoft.Dnx.DesignTimeHost/ApplicationContext.cs
@@ -5,19 +5,18 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
-using System.Reflection;
 using System.Runtime.Versioning;
 using System.Threading;
+using Microsoft.Dnx.Compilation;
+using Microsoft.Dnx.Compilation.Caching;
+using Microsoft.Dnx.Compilation.CSharp;
+using Microsoft.Dnx.Compilation.DesignTime;
 using Microsoft.Dnx.DesignTimeHost.Models;
 using Microsoft.Dnx.DesignTimeHost.Models.IncomingMessages;
 using Microsoft.Dnx.DesignTimeHost.Models.OutgoingMessages;
 using Microsoft.Dnx.Runtime;
-using Microsoft.Dnx.Runtime.Caching;
 using Microsoft.Dnx.Runtime.Common.Impl;
-using Microsoft.Dnx.Compilation;
-using Microsoft.Dnx.Compilation.DesignTime;
 using Microsoft.Dnx.Runtime.Infrastructure;
-using Microsoft.Dnx.Compilation.CSharp;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 using NuGet;
