+++ src/Microsoft.Dnx.DesignTimeHost/Program.cs
@@ -8,9 +8,9 @@ using System.Linq;
 using System.Net;
 using System.Net.Sockets;
 using System.Threading.Tasks;
+using Microsoft.Dnx.Compilation.Caching;
 using Microsoft.Dnx.DesignTimeHost.Models;
 using Microsoft.Dnx.Runtime;
-using Microsoft.Dnx.Runtime.Caching;
 using Microsoft.Dnx.Runtime.Common.DependencyInjection;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
