+++ src/Microsoft.Dnx.Runtime/Compilation/CompositeLibraryExportProvider.cs
