+++ src/Microsoft.Dnx.Runtime/Compilation/Resources/CompositeResourceProvider.cs
