+++ src/Microsoft.Dnx.Runtime/Compilation/Resources/CreateCSharpManifestResourceName.cs
