+++ src/Microsoft.Dnx.Runtime/Compilation/Resources/EmbeddedResourceProvider.cs
