+++ src/Microsoft.Dnx.Runtime/Compilation/Resources/IResourceProvider.cs
