+++ src/Microsoft.Dnx.Runtime/Compilation/Resources/ResourcePathUtility.cs
