+++ src/Microsoft.Dnx.Runtime/Compilation/Resources/ResxResourceProvider.cs
