+++ src/Microsoft.Dnx.Tooling/List/DependencyListOperation.cs
@@ -3,9 +3,9 @@
 
 using System.Linq;
 using System.Runtime.Versioning;
-using Microsoft.Dnx.Tooling.Algorithms;
+using Microsoft.Dnx.Compilation.Caching;
 using Microsoft.Dnx.Runtime;
-using Microsoft.Dnx.Runtime.Caching;
+using Microsoft.Dnx.Tooling.Algorithms;
 using NuGet;
 
 namespace Microsoft.Dnx.Tooling.List
